requirements.txt
# Core dependencies
Flask>=2.2.5
Flask-SQLAlchemy>=3.0.5
SQLAlchemy>=2.0.41
Werkzeug>=2.3.7
requests>=2.31.0

# Telegram Bot
python-telegram-bot>=20.7

# Utilities
python-dotenv>=1.0.1
pytz>=2024.1

# Game-specific
Pillow>=10.3.0
numpy>=1.26.4
pygame>=2.5.2

# Production (optional)
gunicorn>=21.2.0

# Scheduler
APScheduler>=3.10.4

# Misc dependencies
certifi>=2024.2.2
charset-normalizer>=3.3.2
click>=8.1.7
idna>=3.7
importlib-metadata>=7.1.0
itsdangerous>=2.1.2
Jinja2>=3.1.3
MarkupSafe>=2.1.5
typing_extensions>=4.11.0
urllib3>=2.2.1
zipp>=3.18.1

# Visualization
matplotlib>=3.8.4

# Testing
pytest>=8.2.0
pytest-cov>=5.0.0
pytest-mock>=3.14.0

# Development
black>=24.4.2
flake8>=7.0.0

# Additional dependencies
Flask-Login>=0.6.3
Flask-WTF>=1.2.1
email-validator>=2.1.1
flask-migrate>=4.0.5
chapa>=0.1.2
